feat: pivot harness to antigravity (agy); validate pr/issue-substrate hypothesis
Harness pivot (Gemini CLI -> Antigravity):
- Gemini CLI's individual free tier was decommissioned (IneligibleTierError),
  so the primary dogfood + eval harness is now Antigravity (agy): agentic,
  multi-model, and able to import Gemini/Claude plugins.
- Renamed packages/harnesses/gemini-cli -> packages/harnesses/antigravity
  (@engram/harness-antigravity); shell-wrapper now wraps `agy -p`.
- companion: harness name gemini-cli -> antigravity (GEMINI.md `gemini` entry
  kept; Antigravity reads GEMINI.md too).
- Eval runner uses fixture cli_flags (not hardcoded -p) and drops dead tmpfile
  code; fixtures k8s + engram-mcp-decision now target `agy`.
- Antigravity exposes no documented force-injection hook, so delivery leans on
  the shell-wrapper / plugin-import / the gated MCP surface (ADR-009) — which
  reinforces ADR-009. Active plans (near-term-plan W3, direction-2026-h2 P2,
  CLAUDE.md, STATUS) annotated with the pivot. Historical experiment dirs left
  as-is.

PR/issue-substrate validation (the corrected hypothesis from P1):
- New fixture engram-pr-rationale.yaml: substrate = engram git + source +
  GitHub PRs/issues (288 episodes). The rationale (issue #277's beads-inspired
  `engram ready` / `engram observe` proposals) lives ONLY in the issue thread,
  not the tree.
- materialize.sh: builds the bare checkout as a `.git`-LESS `git archive` (a
  worktree's .git let agy follow back to the real repo and leak the answer key
  — caught and fixed), and adds GitHub enrichment.
- VERDICT (experiments/pr-issue-substrate/decision.md): CONFIRMED on two
  questions. bare (full file+code, no issue tracker) does not know #277 exists
  and FABRICATES answers (`engram brief repo`; a format-spec `engram relate`);
  with_pack surfaces #277 and answers correctly (`engram ready` + beads;
  `engram observe` with the exact example). The differentiation the MCP fixture
  could not show — because there the answer was co-located in a tree file.
- Honest caveat: retrieval precision gates the win — the pack surfaces #277 for
  natural phrasings sharing its vocabulary but not for a "reconcile lease/claim"
  phrasing. Improving discussion-retrieval precision is now the top next
  investment.

1622 tests pass; lint clean; build green.
