+++ packages/engram-cli/src/templates/companion/overrides.ts
@@ -11,7 +11,7 @@ export type HarnessName =
   | "claude-code"
   | "cursor"
   | "gemini"
-  | "gemini-cli";
+  | "antigravity";
 
 export const HARNESS_OVERRIDES: Record<HarnessName, string> = {
   generic: `\
@@ -87,26 +87,29 @@ engram companion --harness gemini >> GEMINI.md
 \`\`\`
 `,
 
-  "gemini-cli": `\
-### How to invoke engram context with Gemini CLI (shell-wrapper)
+  antigravity: `\
+### How to invoke engram context with Antigravity (agy, shell-wrapper)
 
-Add the shell-wrapper snippet to your shell profile so every \`gemini\` invocation
-automatically prepends the engram context pack when a \`.engram\` database is present:
+Antigravity (the \`agy\` CLI) has no documented force-injection per-prompt hook, so
+add the shell-wrapper snippet to your shell profile. Every \`agy\` print-mode
+invocation then automatically prepends the engram context pack when a \`.engram\`
+database is present:
 
 \`\`\`sh
-engram companion --harness gemini-cli >> ~/.bashrc
+engram companion --harness antigravity >> ~/.bashrc
 # or for zsh:
-engram companion --harness gemini-cli >> ~/.zshrc
+engram companion --harness antigravity >> ~/.zshrc
 \`\`\`
 
 Reload your shell (\`source ~/.bashrc\`) or start a new session after appending.
 
 The wrapper calls \`engram context\` before each prompt and prepends the pack if a
 \`.engram\` database exists in the current directory. If no \`.engram\` is found, the
-wrapper passes through to \`gemini\` unchanged.
+wrapper passes through to \`agy\` unchanged.
 
-> **Note:** The native \`@engram/harness-gemini-cli\` extension package is currently
-> workspace-only (not published). External users should use the shell-wrapper above.
-> Native extension support will be documented once the Gemini CLI plugin API stabilises.
+> **Note:** Antigravity can also import Gemini/Claude plugins via \`agy plugin
+> import\`, and the gated single-endpoint MCP server (ADR-009) is the sanctioned
+> delivery path for the no-hook case. The shell-wrapper is the zero-setup
+> fallback that works today.
 `,
 };
