+++ packages/engram-core/test/fixtures/eval/k8s-sidecar-containers-753.yaml
@@ -40,10 +40,10 @@ ingest:
 
 evaluation:
   model:
-    provider: gemini
-    model_id: gemini-2.5-pro
-    cli_command: gemini
-    cli_flags: ["-p"]
+    provider: antigravity
+    model_id: gemini-3.1-pro-high
+    cli_command: agy
+    cli_flags: ["--dangerously-skip-permissions", "-p"]
   conditions:
     - id: bare
       description: Agent answers from raw file search alone (no engram pack).
