+++ packages/harnesses/antigravity/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@engram/harness-gemini-cli",
+  "name": "@engram/harness-antigravity",
   "version": "0.1.0",
   "private": true,
   "type": "module",
